// Website/src/Components/Types/User.ts
export type User = {
  id: number;
  name?: string;
  firstName?: string;
  lastName?: string;
  email: string;
  role?: string;
  phoneNumber?: string;
  country?: string;
  city?: string;
  postcode?: string;
  streetName?: string;
  streetNumber?: string;
  positiveSellerCount?: number;
  negativeSellerCount?: number;
  createdAt?: string;
  token?: string;
};